requirements.txt
annotated-types==0.7.0
anyio==4.9.0
beautifulsoup4==4.13.4
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
dotenv==0.9.9
duckduckgo_search==8.0.4
google-api-core==2.25.0
google-auth==2.40.3
google-genai==1.19.0
googleapis-common-protos==1.70.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
lxml==5.4.0
pillow==11.2.1
primp==0.15.0
proto-plus==1.26.1
protobuf==6.31.1
pyaes==1.6.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.5
pydantic_core==2.33.2
python-dotenv==1.1.0
requests==2.32.3
rsa==4.9.1
sniffio==1.3.1
soupsieve==2.7
Telethon==1.40.0
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.4.0
websockets==15.0.1
yt-dlp==2025.6.9
